Include location data sent to Bot
This PR includes locations sharing from the user to the bot.

The workflow is when clicking 'Send my current location' a new message
is sent to the bot with latitude and longitude accessed via
`my_message.location`. To get the coordinates in an array, this can
be accessed via `my_message.location.full`. Original api discussion
[here](https://core.telegram.org/bots/api#location).

This PR does *not* implement sharing of live locations, which are
[implemented in a wholly different manner](https://telegram.org/blog/live-locations) by nesting
the information within `edited_message`. To include live locations we
would also need to bring in the `edited_message` functionality. I will
be approaching this with my next PR.